style: Rename trip_planner to trip-planner
Align naming with add-data, feature-table, save-map.
