+++ projects/hslayers/src/components/trip-planner/layer-selector.component.ts
