+++ projects/hslayers/src/components/trip-planner/public-api.ts
