+++ projects/hslayers/src/components/trip-planner/trip-planner.service.ts
@@ -20,10 +20,10 @@ import {HsEventBusService} from '../core/event-bus.service';
 import {HsLanguageService} from '../language/language.service';
 import {HsLayerUtilsService} from '../utils/layer-utils.service';
 import {HsLayoutService} from '../layout/layout.service';
-import {HsMapService} from './../map/map.service';
-import {HsShareUrlService} from './../permalink/share-url.service';
+import {HsMapService} from '../map/map.service';
+import {HsShareUrlService} from '../permalink/share-url.service';
 import {HsToastService} from '../layout/toast/toast.service';
-import {HsUtilsService} from './../utils/utils.service';
+import {HsUtilsService} from '../utils/utils.service';
 import {getHighlighted} from '../../common/feature-extensions';
 import {getTitle, setTitle} from '../../common/layer-extensions';
